Change format
We found in the service we do the exact same check again. Yeahhh!

Which means we have another place we are specifying the maximum bytes. And this time, it has been expressed in a different format. Yeahhhh!

Anyway, we actually prefer the way it has been expressed in the API service as it doesn't require sticking the value into Google. So, we're updating what we are doing here to use the API format. We'll create a change in the API to get the max bytes in line.