Make charge version max upload consistent (#2179)
We've spotted that the charge version upload endpoint route specifies `5e+7` for `maxBytes`. Quick conversion means this is 50MB. However, the UI page says the max upload is 20MB.

We don't like inconsistency! So, this updates the info in the route config to match what we are stating on the upload page.

We also found in the service we do the exact same check again. This time, it was expressed in a different format. Yeahhhh!

We prefer the way it has been expressed in the API service (it doesn't require sticking the value into Google). So, we're also updating what we are doing here to use the API format. We created [a change in the API](https://github.com/DEFRA/water-abstraction-service/pull/1762) to get the max bytes in line.

Finally, to be consistent all-round we also update the format used in the bulk returns endpoint.