Update Redis config
One of the places NODE_ENV is commonly used in the apps is the Redis config. But we've also spotted that the config is not consistent across the apps.

So, as we have been going through them sorting out NODE_ENV we've also been brining the redis config into line. We've used how things are done in the water-abstraction-service as our template.