Update last wild usages of NODE_ENV
We have left these changes separate because we have not been able to switch to requiring `config.js`. This is because these are shared files but we are in a project that hosts 2 apps with separate config files.

So, as we don't know which one to require we're left sticking with `process.env` directly.