Correctly set NODE_ENV (#2190)
https://github.com/DEFRA/water-abstraction-team/issues/43

We are currently using `NODE_ENV` incorrectly; we set it to be the name of the specific environment (eg. `local`, `qa` etc.) when it should be more like the _type_ of environment (e.g. is it a dev environment, a production environment, etc.)

This means that tools like pm2 aren't correctly picking up that we are running in a dev-like or prod-like environment and configuring accordingly.

We, therefore, change how we use `NODE_ENV` to be the type of environment, with a separate `ENVIRONMENT` env var to set the actual environment (eg. `tst`, `pre` etc.)

** Notes

- Drop requiring dotenv directly
  We've not needed it in other config.js files and we're pretty certain requiring it in the root index.js is all you need.

- Update Redis config
  One of the places NODE_ENV is commonly used in the apps is the Redis config. But we've also spotted that the config is not consistent across the apps.

  So, as we have been going through them sorting out NODE_ENV we've also been bringing the redis config into line. We've used how things are done in the water-abstraction-service as our template.

- Update wild usages of NODE_ENV
  In some places, we have not been able to switch to requiring `config.js` and using the properties we've added. This is because they are shared files but we are in a project that hosts 2 apps with separate config files.

  So, as we don't know which one to require we're left sticking with `process.env` directly.