Fix acceptance-tests proxy for all UIs
We thought we were blocked from being able to run our acceptance tests against our AWS environments due to the limitation that they depend on making calls to the [water-abstraction-service](https://github.com/DEFRA/water-abstraction). This is because unlike locally, only the UI is exposed in our AWS environments. A pass-through or 'proxy' would be needed to allow calls to be sent via the UI to the service API.

It seems the previous team came to the same conclusion because we found `/acceptance-tests/{tail*}` in the routes for both the internal and external app!

However, even though it's listed in both it only works in the external app instance. This change fixes that along with some tidying up of the implementation.