Delete the existing shared plugin
In appears to have originally been built as a shared plugin to make it accesible from both the internal and external app.

Part of fixing and simplifying things is accepting that we only ever need it available in the internal app, and don't need it exposed externally. So, we delete the existing plugin code and move the shared service to the internal `lib/` folder.