Add it to feature toggles
We control other aspects of the app via feature toggles, for example, the ability to 'Delete all billing data'. So, it makes sense to us to include control access to this feature using a feature toggle. That we we can see it listed with the others.

We still only enable the endpoint if the environment is not production.