Add back the feature under internal/modules
We add back in a route file and controller for the `/acceptance-tests` endpoint. It's doing the same thing as before only we hope things are a little clearer.

Our new feature toggle controls whether the endpoint gets added.

If it does it calls our new controller which still uses `AcceptanceTestsProxyService` to do all the work. But this fact is no longer hidden away in some magic method we've added to the request object.

We could have referenced the service in the way a number of other modules do, by requiring `lib/connectors/services` and then using `services.water.acceptanceTestsProxy`. But again that's code trying to be helpful but instead adding complexity and hiding too much of the implementation.

So, we make it clearer what we're using and how by doing it directly in the controller.