Add new /health/info endpoint
https://github.com/DEFRA/water-abstraction-team/issues/54

This is in support of the new service-status page we're building in the [water-abstraction-system repo](https://github.com/DEFRA/water-abstraction-system).

We want to be able to show both the version and commit hash for each service in that page. The existing `/status` page does include the `package.json` version. But that endpoint is also used for the AWS load balancer health check so is heavily hit. We don't want to make it do more work and get the commit hash as well.

Instead, we add a new `/health/` path and a new `/info` endpoint to return both the version and commit hash.