Refactor routes and re-write tests
The previous version might have worked. But the fact the test had become so complex and needed special exceptions was a critical indicator that something was wrong.

So, rather than have the logic in the routes we chose to keep them simple; these are the routes. We moved the logic of whether to include them to the place where they get applied. Doing that cleans up the tests.

Admittedly, what is left is not really checking much. But it does at least serve as documentation.