Create system proxy endpoint (#2229)
https://eaflood.atlassian.net/browse/WATER-3787
https://github.com/DEFRA/water-abstraction-team/issues/54

Going forward, we intend all new functionality to be built in the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) repo. We even intend to _try_ and migrate any existing functionality we touch. In a perfect world, we'll end up with all the existing repos becoming defunct and a clean, performant service all in one repo.

To get there though, we still need to work with the existing services and the existing environments. In those, the only apps exposed are the internal and eternal UI which are based on this repo.

So, for any requests we want the water-abstraction-system to handle we need to 'proxy' them via the water-abstraction-ui. So, this change adds the ability for the UI to proxy all requests sent to `/system` to the water-abstraction-system app behind the scenes.

One of the challenges we hit was assets. We realised when we requested a view, assets would also be requested. So, included in the change is the ability for the UI to proxy asset requests for the system as well.