Fix verification code appearing in production environment
https://eaflood.atlassian.net/browse/WATER-3816

When registering a new account, the end of the process triggers a verification code being sent to the customer's postal address. In our test environments, we display this code on the page so the user can continue without needing to receive the code by post. However it's been reported that this code recently appeared to a user in the production environment. We therefore fix this defect.