Add feature flag to fix verification code appearing in production environment (#2234)
https://eaflood.atlassian.net/browse/WATER-3816

When registering a new account, the end of the process triggers a verification code being sent to the customer's postal address. In our test environments, we display this code on the page so the user can continue without needing to receive the code by post. However it's been reported that this code recently appeared to a user in the production environment. We therefore fix this defect.

We do this by creating a new feature flag, defined in `.env` as `SHOW_VERIFICATION_CODE_FEATURE`, which displays the code when set to `true` but with an additional check so it is only displayed in non-prod environments. We then swap out the previous use of the `TEST_MODE` env var for our new feature flag.