Fix request.logout() is undefined error (#2235)
https://eaflood.atlassian.net/browse/WATER-3815

An external user reported seeing our SilverLine WAF error page. web-ops tracked this down to `Illegal HTTP status in response`. Basically, the service threw an unhandled error which caused it to return a http 500 response code. Responses like this are blocked because they typically leak information about the site.

We tracked down the issue being that the app was rejecting the request due to an invalid CSRF token (why, we'll never know!) That was then causing the error plugin to fire. In the event of a CRSF issue the error plugin will immediately log the user out using `request.logOut()`.

The problem is that function is attached to the Hapi request object in the auth plugin. It is also attached to the Hapi `onPreHandler` extension point, but registered _after_ the CSRF plugin. Hapi runs plugins in the order they are registered, which is why we get a situation where the CRSF logic and subsequent error handling is running before the auth plugin has had a chance to decorate the request object.

To fix this bug we have moved where the plugin runs to be higher up in the hapi request lifecycle.