Fix system proxy endpoint forwarding query params
https://github.com/DEFRA/water-abstraction-team/issues/54
https://eaflood.atlassian.net/browse/WATER-3787

In a previous change we [Create system proxy endpoint](https://github.com/DEFRA/water-abstraction-ui/pull/2229). We need this to be able to send requests to the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) when using our AWS environments.

That all seemed to work, until we changed an endpoint in **water-absraction-system** to require a query param to be passed through. Our proxy isn't doing this so now **water-abstraction-system** errors because of the missing param.

This change fixes the issue.