Add h2o2 plugin package
[h2o2](https://github.com/hapijs/h2o2) "adds proxying functionality to a hapi server."

When we first implemented our proxy, we did it ourselves. This issue with the query params would require us to continue to expand that functionality, and we haven't even got to `POST`, `PUT` and `DELETE` requests yet!

So, before we went down that route we investigated what options were available and found this plugin (and there is only one, this being Hapi 😩). The documentation and current examples all direct you to use this if you need Hapi to act like a reverse/pass-through proxy server.

We should have done this from the outset, but we've at least stopped before we expanded our own functionality further.