Add logic to trigger SROC supplementary from PRESROC (#2247)
https://eaflood.atlassian.net/browse/WATER-3828

When presroc supplementary billing has completed, we want to trigger sroc supplementary billing. To do this we update the existing billing routing so that when the current batch has reached `ready` status, we check to see if it's a presroc supplementary batch so we can trigger the creation of an sroc supplementary batch if needed, and continue the billing flow using this new batch so the user once again sees the spinner page until it has completed.

At present we don't have the ability to create sroc supplementary batches so we simply create a dummy batch with its status  already set to `ready`. 

This is all hidden behind a feature flag `TRIGGER_SROC_SUPPLEMENTARY`.