Rework connector as service client
We realised that the API client we'd created is more typically used to connect via `hapi-pg-rest-api` (although typically we found other instances where this isn't the case!). We therefore rework it as a service client. We also move it to the `internal` folder as we don't currently expect it to be shared.