Remove external route and just keep internal
Didn't realise we had exactly the same endpoint doing exactly the same thing in both the internal and external app. We test the UI as a single service so there is no point in this duplication. So, we remove the external version and just keep the internal one going forward.