Make /notifications/last available in non-prod (#2267)
Currently, the external endpoint `/notifications/last?email=acceptance-test.external@example.com` is not available in our non-production environments. This is because it still relies on the old `TEST_MODE` env var rather than our new `isProduction` logic in `config.js`. This means we can't make the acceptance tests work against our AWS environments and will be stuck running them locally.

This change updates the logic. The endpoint will still only be available in non-production environments. But we'll use our updated logic to determine this. We also reduce the maintenance overhead by only exposing it in the internal app (the logic was duplicated in both).