Label empty bill runs as `EMPTY`
https://eaflood.atlassian.net/browse/WATER-3828

When you run a bill run for a region and the service finds no applicable licences that need billing it directs you to an error page. The page tells you the reason is that the bill run is empty, so this fact is in the data somewhere!

What we and the users would ideally like to see is the bill run labelled as `EMPTY` in the view bill runs screen. As we'll have 2 distinct supplementary bill runs being generated each time it's important we can start distinguishing between those that have truly errored, and those where there was nothing to do.