Add new error page for errored bill runs
https://eaflood.atlassian.net/browse/WATER-3882

When working on WATER-3833 it was tested what would happen if an error occurred.

If this happens in the legacy code the `billing_batch` has its `status` changed to error. In the UI it still gets listed with an **ERROR** label. In our new SROC supplementary bill run code though we weren't creating the record.

To ensure parity between the legacy and new code, we did some testing of what the legacy code actually does. What we found was _if_ it manages to throw and handle an error, it does set the `status` to 'error' and in some cases it also sets an `error_code` against the bill run. But though the team thought there was an error page to display this information what actually happens is the service forwards the errored bill runs to `/billing/batch/${id}/processing`, which then 'helpfully' throws a HTTP `500`

```javascript
  // src/internal/modules/billing/controllers/bill-run.js - getBillingBatchProcessing()
  // ...

  // Render error page if batch has errored
  if (batch.status === 'error') {
    return Boom.badImplementation('Billing batch error')
  }

  // ...
```

The error page the comment is referring to is the generic 'Something's gone wrong page'. But the user will never see this because it also returns a HTTP `500` code, which our WAF blocks for good reasons (never send a 500 to the user!)

Why is this an issue now? We're implementing a process where the PRESROC and SROC bill runs are created in parallel, but processed in sequence. If an error happens during the processing of the SROC bill run that`s fine because it is at the forefront. But if an error happens during creation, the PRESROC is at the forefront. We can`t redirect to an error page for something happening in the background. All the user will see is the SROC bill run listed as **ERROR** with no other details.

To improve their experience, avoid needless ☎️ to the team, and stop users from seeing the Firewall error page (at least in this context) we're adding an error page, that errored bill runs can link through to.