Remove the isErrorRoutesIncluded option
It was redundant in the first place but now it definately is.

Also some initial tidy up of the unit tests for routing. We move tests of 'empty' status up a level to its own block. We also use more realistic looking IDs for our test IDs.

Gonna do some more tidying in the next change.