Argh! Should not have touched back link
Realised that the back link in `error` and `empty` pages is hard set. I started playing with the back link options because I was worried about these pages trying to send a user back to a page in the create bill run process.

Having clocked the link _always_ takes the user to the list bill runs page realised this is _exactly_ what we want and I should have left things alone!