Remove unnecessary tear-downs in acceptance tests (#2283)
In all of our acceptance tests, we use `before()` and `after()` [hooks](https://docs.cypress.io/guides/core-concepts/writing-and-organizing-tests#Hooks) to set the state of the app.

In the `before()` hook we'll call the `/tear-down` endpoint to delete the existing test data. We'll then follow it up with a call to set up the test data we need, for example, `/barebones`.

It's _extremely_ slow. But it's all we've got at the moment.

For some reason though, all our tests also have an `after()` hook that calls `/tear-down` again. We always call it in our `before()` hook. So, it's not adding any value. Add to that [Cypress best practices](https://docs.cypress.io/guides/references/best-practices#Using-after-or-afterEach-hooks) specifically state you should not do this!

On the basis it's pointless, against best practices, and slows the tests down this change removes the `after()` hooks.