Disable video in Cypress
By default, for each test Cypress will record and then compress a video. These might be handy for sharing to those outside the team. But we never do! They can also help with debugging. But when we need to we just run the problem spec in the GUI.

So, it's another thing that are slowing the tests down and not adding any benefit.

This change disables videos by default.