Make default env vars for Cypress consistent (#2285)
Spotted in our Cypress acceptance tests that one of the URIs is defined differently from the others. No need for this. So, this is a quick change to make them consistent.

---

Also, added a `use strict` and denoted the first param as unused. We touched the file so did a little tidying whilst in there.