Fix Cypress get email link commands
We hadn't spotted that when me made a change recently to [allow getting the last notification sent in our non-prod environments](https://github.com/DEFRA/water-abstraction-ui/pull/2267) that we broke the Cypress acceptance tests that depend on following these links, for example, password reset.

Simply put, the commands were premised on the endpoint being available in both the internal and external version of the UI. Now it's only available in the internal version, we can tweak the commands and fix the issue.