Always generate the Cypress test reports
The `package.json` scripts that run the Cypress tests automatically call the `test:cypress-prepare-report` automatically when finished. But because of the `&&` it only does it if all the tests pass.

We know we have tests that don't, mainly because we are still trying to figure out what they were testing! But this means Cypress never finishes with a 0 exit code, which means preparing the test report is never called.

This tweak ensures the we always generate the test report, even if Cypress had failed tests.