Tidy up cleaning of Cypress reports
Before you generate a new report you need to clean our the artifacts from the previous run. So, the previous team added a script to do this.

But they also hooked it in to `pretest:`, which means whenever tests are run, including our unit tests, we run this command. We only need to run it when calling `cypress run`, as that's the only time its generating report output.

To make this clear, we renamed the script label to follow the convention for other Cypress scripts. We also removed the `pretest:` script.