Always generate the Cypress test reports (#2289)
The `package.json` scripts that run the Cypress tests automatically call the `test:cypress-prepare-report` automatically when finished. But because of the `&&` it only does it if all the tests pass.

We know we have tests that don't, mainly because we are still trying to figure out what they were testing! But this means Cypress never finishes with a 0 exit code, which means preparing the test report is never called.

This tweak ensures that we always generate the test report, even if Cypress had failed tests.

---

Whilst making this change we also took the opportunity to do a bit of housekeeping. We've removed commands we never use. We've also sorted out the cleaning of the Cypress reports folder.

Before you generate a new report you need to clean out the artefacts from the previous run. So, the previous team added a script to do this.

But they also hooked it into `pretest:`, which means whenever tests are run, including our unit tests, we run this command. We only need to run it when calling `cypress run`, as that's the only time it's generating report output.

So, we renamed the script label to follow the convention for other Cypress scripts. We also removed the `pretest:` script.