Implement fix
Quite simply we can take the change previously made in [Bugfix/water 3348](https://github.com/DEFRA/water-abstraction-ui/pull/1943) and apply it to the `getReviewChargeInformation()` controller.

We also spotted that the `postReviewChargeInformation()` controller does exactly the same thing again if the form submitted is invalid. So, we needed to make the change there as well.