Fix licence holder display bug in check charge info (#2300)
https://eaflood.atlassian.net/browse/WATER-3905

The business has raised an issue with the 'Check charge information' view. The scenario they outlined was

- Licence **AN/030/0008/006** was originally issued to _Cemex Materials Ltd_ with effect from 14/03/2022
- It was then transferred to _Breedon Trading Limited_ with effect from 27/06/2022
- Within the period 14/03/2022 - 26/06/2022 they would expect to see _Cemex Materials Ltd_ as the licence holder
- From 27/06/2022 we would expect to see _Breedon Trading Limited_ as the licence holder

This reminded someone that we'd had a similar issue previously. They found WATER-3348 which is linked to [Bugfix/water 3348](https://github.com/DEFRA/water-abstraction-ui/pull/1943). It seems that fixed the bug for the charge info view page (`/licences/{licenceId}/charge-in
formation/{chargeVersionId}/view`). But though the controller for `/licences/{licenceId}/charge-information/{chargeVersionWorkflowId}/review` is directly below it, it didn't get fixed at the same time. 🙁

Whilst digging into this we also found the same issue with the `/licences/{licenceId}/charge-information/check` screen.

So, this fix applies what was done to `getViewChargeInformation()` to `getReviewChargeInformation()` and `getCheckData()` in `src/internal/modules/charge-information/controllers/charge-information.js`.