Apply fix
For reference, the issue found was that `documentRoles` contained 2 entries with the flag 'licenceHolder'.

- **Cemex**: start date `2022-03-14`, end date `2022-06-26`
- **Breedon**: start date `2022-06-27`, end date `2028-03-31`

Setting the charge version start date (`draftChargeInformation.dateRange.startDate`) to be `2022-06-25` or `2022-06-27` would both work because based on the filter being applied either **Cemex** or **Breedon** would be matched. But when `2022-06-26` was selected **Cemex** would match the start date clause, but nothing would match the end date.

A simple tweak to use [isSameOrAfter()](https://momentjs.com/docs/#/query/is-same-or-after/) saved the day! Now the **Cemex** entry matches both clauses, and that is the licence holder the user expects to see.