Remove axios from acceptance tests (#2314)
We'd already spotted that for some reason _another_ HTTP request package had been added to the repo just for use in Cypress. This was even more confusing because Cypress is itself a framework for sending HTTP requests!

Ideally, we'd like to run the current acceptance tests from the same repo we download as part of our Docker base local dev environment. You have to run the tests outside of the environment though because of the dependence on browsers being available. But because of the inclusion of [axios](https://github.com/axios/axios) as part of the dependencies any tests that rely on the custom commands that use **axios** fail.

So, this change removes **axios** and replaces what it was doing with native Cypress commands. It helps us remove a redundant package (that is also out of date) and allows us to avoid having to maintain 2 copies of this repo on our machines.